Refactor storage code
a.) Break storage code into smaller readable units with each structure
and operations around it in the same file
b.) Remove commented code
c.) Rename storage.go, storage.go to
block_devices.go
block_devices_ops.go respectively
d.) Remove unused function FreeSpace
