+++ storage/block_devices.go
@@ -8,8 +8,6 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
-	"math"
 	"path/filepath"
 	"regexp"
 	"strconv"
@@ -20,18 +18,6 @@ import (
 	"github.com/clearlinux/clr-installer/log"
 	"github.com/clearlinux/clr-installer/utils"
 )
-
-// PartedPartition hold partition information
-// Number 0 and FileSystem "free" are free spaces
-type PartedPartition struct {
-	Number     uint64 // partition number 0 indicates free space
-	Start      uint64 // starting byte location
-	End        uint64 // ending byte location
-	Size       uint64 // size in bytes
-	FileSystem string // file system Type
-	Name       string // partition name
-	Flags      string // flags for partition
-}
 
 // A BlockDevice describes a block device and its partitions
 type BlockDevice struct {
@@ -61,25 +47,6 @@ type BlockDevice struct {
 	partition       uint64             // Assigned partition for media - can't set until after mkpart
 	PartTable       []*PartedPartition // Existing Disk partition table from parted
 	removedParts    []uint64           // List of manually removed partitions
-}
-
-// Version used for reading and writing YAML
-type blockDeviceYAMLMarshal struct {
-	Name            string         `yaml:"name,omitempty"`
-	Model           string         `yaml:"model,omitempty"`
-	MajorMinor      string         `yaml:"majMin,omitempty"`
-	FsType          string         `yaml:"fstype,omitempty"`
-	UUID            string         `yaml:"uuid,omitempty"`
-	Serial          string         `yaml:"serial,omitempty"`
-	MountPoint      string         `yaml:"mountpoint,omitempty"`
-	Label           string         `yaml:"label,omitempty"`
-	Size            string         `yaml:"size,omitempty"`
-	ReadOnly        string         `yaml:"ro,omitempty"`
-	RemovableDevice string         `yaml:"rm,omitempty"`
-	Type            string         `yaml:"type,omitempty"`
-	State           string         `yaml:"state,omitempty"`
-	Children        []*BlockDevice `yaml:"children,omitempty"`
-	Options         string         `yaml:"options,omitempty"`
 }
 
 // BlockDeviceState is the representation of a block device state (live, running, etc)
@@ -202,7 +169,6 @@ const (
 var (
 	avBlockDevices      []*BlockDevice
 	lsblkBinary         = "lsblk"
-	storageExp          = regexp.MustCompile(`^([0-9]*(\.)?[0-9]*)([bkmgtp]{1}){0,1}$`)
 	devNameSuffixExp    = regexp.MustCompile(`([0-9]*)$`)
 	blockDeviceStateMap = map[BlockDeviceState]string{
 		BlockDeviceStateRunning:   "running",
@@ -367,21 +333,6 @@ func (bd *BlockDevice) findFree(size uint64) *PartedPartition {
 	return freePart
 }
 
-// Clone creates a copies a PartedPartition
-func (part *PartedPartition) Clone() *PartedPartition {
-	clone := &PartedPartition{
-		Number:     part.Number,
-		Start:      part.Start,
-		End:        part.End,
-		Size:       part.Size,
-		FileSystem: part.FileSystem,
-		Name:       part.Name,
-		Flags:      part.Flags,
-	}
-
-	return clone
-}
-
 // Clone creates a copies a BlockDevice and its children
 func (bd *BlockDevice) Clone() *BlockDevice {
 	clone := &BlockDevice{
@@ -498,94 +449,6 @@ func (bd *BlockDevice) AddChild(child *BlockDevice) {
 		}
 	}
 	log.Debug("AddChild: child.Name is %q", child.Name)
-}
-
-// HumanReadableSizeWithUnitAndPrecision converts the size representation in bytes to the
-// closest human readable format i.e 10M, 1G, 2T etc with a forced unit and precision
-func HumanReadableSizeWithUnitAndPrecision(size uint64, unit string, precision int) (string, error) {
-	unit = strings.ToUpper(unit)
-
-	if size == 0 {
-		return fmt.Sprintf("0"), nil
-	}
-
-	sizes := []struct {
-		unit      string
-		mask      float64
-		precision int
-	}{
-		{"P", 1.0 * 1000.0 * 1000.0 * 1000.0 * 1000.0 * 1000.0, 5},
-		{"T", 1.0 * 1000.0 * 1000.0 * 1000.0 * 1000.0, 4},
-		{"G", 1.0 * 1000.0 * 1000.0 * 1000.0, 3},
-		{"M", 1.0 * 1000.0 * 1000.0, 2},
-		{"K", 1.0 * 1000.0, 1},
-		{"B", 1.0, 0},
-	}
-
-	value := float64(size)
-	for _, curr := range sizes {
-		csize := value / curr.mask
-
-		// No unit request, use default based on size
-		if unit == "" {
-			if csize < 1 {
-				continue
-			}
-		} else if unit != curr.unit {
-			continue
-		}
-
-		unit = curr.unit
-
-		// No precision request, use default based on size
-		if precision < 0 {
-			precision = curr.precision
-		}
-
-		formatted := strconv.FormatFloat(csize, 'f', precision, 64)
-		// Remove trailing zeroes (and unused decimal)
-		formatted = strings.TrimRight(strings.TrimRight(formatted, "0"), ".")
-		if unit != "" && unit != "B" {
-			formatted += unit
-		}
-
-		return formatted, nil
-	}
-
-	return "", errors.ValidationErrorf("Could not format disk/partition size")
-}
-
-// HumanReadableSizeWithPrecision converts the size representation in bytes to the
-// closest human readable format i.e 10M, 1G, 2T etc with a forced precision
-func HumanReadableSizeWithPrecision(size uint64, precision int) (string, error) {
-	return HumanReadableSizeWithUnitAndPrecision(size, "", precision)
-}
-
-// HumanReadableSizeWithUnit converts the size representation in bytes to the
-// closest human readable format i.e 10M, 1G, 2T etc with a forced unit
-func HumanReadableSizeWithUnit(size uint64, unit string) (string, error) {
-	return HumanReadableSizeWithUnitAndPrecision(size, unit, -1)
-}
-
-// HumanReadableSize converts the size representation in bytes to the closest
-// human readable format i.e 10M, 1G, 2T etc
-func HumanReadableSize(size uint64) (string, error) {
-	return HumanReadableSizeWithUnitAndPrecision(size, "", -1)
-}
-
-// FreeSpace returns the block device available/free space considering the currently
-// configured partition table
-func (bd *BlockDevice) FreeSpace() (uint64, error) {
-	if !utils.IntSliceContains([]int{BlockDeviceTypeDisk, BlockDeviceTypeLoop}, int(bd.Type)) {
-		return 0, errors.Errorf("FreeSpace() must only be called with a disk block device")
-	}
-
-	var total uint64
-	for _, curr := range bd.Children {
-		total = total + curr.Size
-	}
-
-	return bd.Size - total, nil
 }
 
 // HumanReadableSizeWithUnitAndPrecision converts the size representation in bytes to the
@@ -777,6 +640,7 @@ func isBlockDeviceAvailable(blockDevices []*BlockDevice) bool {
 
 	return available
 }
+
 func parseBlockDevicesDescriptor(data []byte) ([]*BlockDevice, error) {
 	root := struct {
 		BlockDevices []*BlockDevice `json:"blockdevices"`
@@ -797,511 +661,6 @@ func parseBlockDevicesDescriptor(data []byte) ([]*BlockDevice, error) {
 	}
 
 	return root.BlockDevices, nil
-}
-
-func getNextStrToken(dec *json.Decoder, name string) (string, error) {
-	t, _ := dec.Token()
-	if t == nil {
-		return "", nil
-	}
-
-	str, valid := t.(string)
-	if !valid {
-		return "", errors.Errorf("\"%s\" token should have a string value", name)
-	}
-
-	return str, nil
-}
-
-func getNextByteToken(dec *json.Decoder, name string) (uint64, error) {
-	var byteSize uint64
-	var err error
-
-	dec.UseNumber()
-	token, _ := dec.Token()
-	if token == nil {
-		return 0, nil
-	}
-
-	switch t := token.(type) {
-	case json.Number:
-		// Is it an unsigned int value (lsblk >= 2.33)
-		var n int64
-
-		n, err = t.Int64()
-		if err != nil {
-			return 0, err
-		}
-
-		byteSize = uint64(n)
-
-	case string:
-		// Is it a string value (lsblk < 2.33)
-
-		str, sValid := token.(string)
-		if !sValid {
-			return 0, errors.Errorf("\"%s\" token is neither an uint64 nor a string value", name)
-		}
-
-		byteSize, err = ParseVolumeSize(str)
-		if err != nil {
-			return 0, err
-		}
-	}
-
-	return byteSize, nil
-}
-
-func getNextBoolToken(dec *json.Decoder, name string) (bool, error) {
-	t, _ := dec.Token()
-	if t == nil {
-		return false, nil
-	}
-
-	// Is it a boolean value (lsblk >= 2.33)
-	b, bValid := t.(bool)
-	if bValid {
-		return b, nil
-	}
-
-	// Is it a string value (lsblk < 2.33)
-	str, sValid := t.(string)
-	if !sValid {
-		return false, errors.Errorf("\"%s\" token is neither a boolean nor a string value", name)
-	}
-
-	if str == "0" {
-		return false, nil
-	} else if str == "1" {
-		return true, nil
-	} else if str == "" {
-		return false, nil
-	}
-
-	return false, errors.Errorf("Unknown ro value: %s", str)
-}
-
-// ParseVolumeSize will parse a string formatted (1M, 10G, 2T) size and return its representation
-// in bytes
-func ParseVolumeSize(str string) (uint64, error) {
-	var size uint64
-
-	str = strings.ToLower(str)
-
-	if !storageExp.MatchString(str) {
-		return strconv.ParseUint(str, 0, 64)
-	}
-
-	unit := storageExp.ReplaceAllString(str, `$3`)
-	fsize, err := strconv.ParseFloat(storageExp.ReplaceAllString(str, `$1`), 64)
-	if err != nil {
-		return 0, errors.Wrap(err)
-	}
-
-	switch unit {
-	case "b":
-		fsize = fsize * (1 << 0)
-	case "k":
-		fsize = fsize * (1 << 10)
-	case "m":
-		fsize = fsize * (1 << 20)
-	case "g":
-		fsize = fsize * (1 << 30)
-	case "t":
-		fsize = fsize * (1 << 40)
-	case "p":
-		fsize = fsize * (1 << 50)
-	}
-
-	size = uint64(math.Round(fsize))
-
-	return size, nil
-}
-
-// UnmarshalJSON decodes a BlockDevice, targeted to integrate with json
-// decoding framework
-func (bd *BlockDevice) UnmarshalJSON(b []byte) error {
-
-	dec := json.NewDecoder(bytes.NewReader(b))
-
-	for {
-		t, err := dec.Token()
-		if err == io.EOF {
-			break
-		}
-
-		str, valid := t.(string)
-		if !valid {
-			continue
-		}
-
-		switch str {
-		case "name":
-			var name string
-
-			name, err = getNextStrToken(dec, "name")
-			if err != nil {
-				return err
-			}
-
-			bd.Name = name
-		case "model":
-			var model string
-
-			model, err = getNextStrToken(dec, "model")
-			if err != nil {
-				return err
-			}
-
-			bd.Model = model
-		case "maj:min":
-			var majMin string
-
-			majMin, err = getNextStrToken(dec, "maj:min")
-			if err != nil {
-				return err
-			}
-
-			bd.MajorMinor = majMin
-		case "size":
-			var size uint64
-
-			size, err = getNextByteToken(dec, "size")
-			if err != nil {
-				return err
-			}
-
-			bd.Size = size
-		case "pttype":
-			var pttype string
-
-			pttype, err = getNextStrToken(dec, "pttype")
-			if err != nil {
-				return err
-			}
-
-			bd.PtType = pttype
-		case "fstype":
-			var fstype string
-
-			fstype, err = getNextStrToken(dec, "fstype")
-			if err != nil {
-				return err
-			}
-
-			bd.FsType = fstype
-		case "uuid":
-			var uuid string
-
-			uuid, err = getNextStrToken(dec, "uuid")
-			if err != nil {
-				return err
-			}
-
-			bd.UUID = uuid
-		case "serial":
-			var serial string
-
-			serial, err = getNextStrToken(dec, "serial")
-			if err != nil {
-				return err
-			}
-
-			bd.Serial = serial
-		case "type":
-			var tp string
-
-			tp, err = getNextStrToken(dec, "type")
-			if err != nil {
-				return err
-			}
-
-			bd.Type, err = parseBlockDeviceType(tp)
-			if err != nil {
-				return err
-			}
-		case "state":
-			var state string
-
-			state, err = getNextStrToken(dec, "state")
-			if err != nil {
-				return err
-			}
-
-			bd.State, err = parseBlockDeviceState(state)
-			if err != nil {
-				return err
-			}
-		case "mountpoint":
-			var mpoint string
-
-			mpoint, err = getNextStrToken(dec, "mountpoint")
-			if err != nil {
-				return err
-			}
-
-			bd.MountPoint = mpoint
-		case "label":
-			var label string
-
-			label, err = getNextStrToken(dec, "label")
-			if err != nil {
-				return err
-			}
-
-			bd.Label = label
-		case "partlabel":
-			var label string
-
-			label, err = getNextStrToken(dec, "partlabel")
-			if err != nil {
-				return err
-			}
-
-			bd.PartitionLabel = label
-		case "ro":
-			bd.ReadOnly, err = getNextBoolToken(dec, "ro")
-			if err != nil {
-				return err
-			}
-		case "rm":
-			bd.RemovableDevice, err = getNextBoolToken(dec, "rm")
-			if err != nil {
-				return err
-			}
-		case "children":
-			bd.Children = []*BlockDevice{}
-			err := dec.Decode(&bd.Children)
-			if err != nil {
-				return errors.Errorf("Invalid \"children\" token: %s", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-// MarshalYAML is the yaml Marshaller implementation
-func (bd *BlockDevice) MarshalYAML() (interface{}, error) {
-
-	var bdm blockDeviceYAMLMarshal
-
-	bdm.Name = bd.Name
-	bdm.Model = bd.Model
-	bdm.MajorMinor = bd.MajorMinor
-	bdm.FsType = bd.FsType
-	bdm.UUID = bd.UUID
-	bdm.Serial = bd.Serial
-	bdm.MountPoint = bd.MountPoint
-	bdm.Label = bd.Label
-	bdm.Size = strconv.FormatUint(bd.Size, 10)
-	bdm.ReadOnly = strconv.FormatBool(bd.ReadOnly)
-	bdm.RemovableDevice = strconv.FormatBool(bd.RemovableDevice)
-	bdm.Type = bd.Type.String()
-	bdm.State = bd.State.String()
-	bdm.Children = bd.Children
-	bdm.Options = bd.Options
-
-	return bdm, nil
-}
-
-// UnmarshalYAML is the yaml Unmarshaller implementation
-func (bd *BlockDevice) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	var unmarshBlockDevice blockDeviceYAMLMarshal
-
-	if err := unmarshal(&unmarshBlockDevice); err != nil {
-		return err
-	}
-
-	// Copy the unmarshaled data
-	bd.Name = unmarshBlockDevice.Name
-	bd.Model = unmarshBlockDevice.Model
-	bd.MajorMinor = unmarshBlockDevice.MajorMinor
-	bd.FsType = unmarshBlockDevice.FsType
-	bd.UUID = unmarshBlockDevice.UUID
-	bd.Serial = unmarshBlockDevice.Serial
-	bd.MountPoint = unmarshBlockDevice.MountPoint
-	bd.Label = unmarshBlockDevice.Label
-	bd.Children = unmarshBlockDevice.Children
-	bd.Options = unmarshBlockDevice.Options
-	// Convert String to Uint64
-	if unmarshBlockDevice.Size != "" {
-		uSize, err := ParseVolumeSize(unmarshBlockDevice.Size)
-		if err != nil {
-			return err
-		}
-		bd.Size = uSize
-	}
-
-	// Map the BlockDeviceType
-	if unmarshBlockDevice.Type != "" {
-		iType, err := parseBlockDeviceType(unmarshBlockDevice.Type)
-		if err != nil {
-			return errors.Errorf("Device: %s: %v", unmarshBlockDevice.Name, err)
-		}
-		if iType < 0 || iType > BlockDeviceTypeUnknown {
-		}
-		bd.Type = iType
-		if iType != BlockDeviceTypeDisk {
-			bd.MakePartition = true
-			bd.FormatPartition = true
-		}
-	}
-
-	// Map the BlockDeviceState
-	if unmarshBlockDevice.State != "" {
-		iState, err := parseBlockDeviceState(unmarshBlockDevice.State)
-		if err != nil {
-			return errors.Errorf("Device: %s: %v", unmarshBlockDevice.Name, err)
-		}
-		bd.State = iState
-	}
-
-	// Map the ReanOnly bool
-	if unmarshBlockDevice.ReadOnly != "" {
-		bReadOnly, err := strconv.ParseBool(unmarshBlockDevice.ReadOnly)
-		if err != nil {
-			return err
-		}
-		bd.ReadOnly = bReadOnly
-	}
-
-	// Map the RemovableDevice bool
-	if unmarshBlockDevice.RemovableDevice != "" {
-		bRemovableDevice, err := strconv.ParseBool(unmarshBlockDevice.RemovableDevice)
-		if err != nil {
-			return err
-		}
-		bd.RemovableDevice = bRemovableDevice
-	}
-
-	return nil
-}
-
-// MaxLabelLength returns the maximum length of a label for
-// the given file system type
-func MaxLabelLength(fstype string) int {
-	var maxLen int
-
-	switch fstype {
-	case "ext2", "ext3", "ext4":
-		maxLen = 16
-	case "swap":
-		maxLen = 15
-	case "xfs":
-		maxLen = 12
-	case "f2fs":
-		maxLen = 512
-	case "btrfs":
-		maxLen = 255
-	case "vfat":
-		maxLen = 11
-	default:
-		maxLen = 11
-		log.Warning("Unknown file system type %s, defaulting to %d character label", fstype, maxLen)
-	}
-
-	return maxLen
-}
-
-// AddBootStandardPartition will add to disk a new standard Boot partition
-func AddBootStandardPartition(disk *BlockDevice) uint64 {
-	freePart := disk.findFree(bootSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            bootSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "vfat",
-		MountPoint:      "/boot",
-		Label:           "boot",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
-
-	return bootSize
-}
-
-// AddSwapStandardPartition will add to disk a new standard Swap partition
-func AddSwapStandardPartition(disk *BlockDevice) uint64 {
-	freePart := disk.findFree(swapSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            swapSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "swap",
-		Label:           "swap",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
-
-	return swapSize
-}
-
-// AddRootStandardPartition will add to disk a new standard Root partition
-func AddRootStandardPartition(disk *BlockDevice, rootSize uint64) {
-	freePart := disk.findFree(rootSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            rootSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "ext4",
-		MountPoint:      "/",
-		Label:           "root",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
-}
-
-// NewStandardPartitions will add to disk a new set of partitions representing a
-// default set of partitions required for an installation
-func NewStandardPartitions(disk *BlockDevice) {
-	disk.Children = nil
-	newFreePart := &PartedPartition{
-		Number:     0,
-		Start:      0,
-		End:        disk.Size,
-		Size:       disk.Size,
-		FileSystem: "free",
-	}
-	disk.PartTable = nil
-	disk.PartTable = append(disk.PartTable, newFreePart)
-
-	rootSize := uint64(disk.Size - bootSize - swapSize)
-
-	freePart := disk.findFree(bootSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            bootSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "vfat",
-		MountPoint:      "/boot",
-		Label:           "boot",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
-
-	freePart = disk.findFree(swapSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            swapSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "swap",
-		Label:           "swap",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
-
-	freePart = disk.findFree(rootSize)
-	disk.AddFromFreePartition(freePart, &BlockDevice{
-		Size:            rootSize,
-		Type:            BlockDeviceTypePart,
-		FsType:          "ext4",
-		MountPoint:      "/",
-		Label:           "root",
-		UserDefined:     true,
-		MakePartition:   true,
-		FormatPartition: true,
-	})
 }
 
 // PartProbe runs partprobe against the block device's file
